New Scripts, Particles and some bug fixing
Added Scripts for every appliance that will be added + their respective particle spawner scripts.

Added new Particles to indicate what appliance is on.

Added footstep particles for the kid.

Adjustment to money per second formula to accommodate for the addition of more appliances.

Fixed a bug in the upgrade menu where after an house upgrade was bough the button to buy the same upgrade would appear after the scene gets reloaded.

Added/Fixed the way the money per second would be calculated after an upgrade was bought. Now all 4 of the house upgrades work as intended.

Added foundation for the Teddy Bear Grenade Tool Upgrade. The player can throw a grenade and something happens after a set amount of delay. Needs Model, Explosion Effect and an adjustment in the kids script for it to target the grenade.

Fixed an issue where the kid would camp the last turned on appliance. Now after everything is on the kid will go to a set destination away from the appliances and stay there until an appliance has been turned off.

Added and hocked up the walking and appliance interacting animations for the kid.